requirements-worker.txt
# Prefect workflow orchestration
prefect==2.14.0

# InfluxDB client
influxdb-client==1.38.0

# Data processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# HTTP requests (for Tibber API)
requests==2.31.0

# Excel support (for data import - optional)
openpyxl==3.1.2

# Plotting (for reporter - optional, can be removed if not needed)
plotly==5.18.0
